requirements.txt
paho-mqtt
openai
requests
